refactor: move KeyStageKeypad to SharedComponents
