+++ src/components/SharedComponents/KeyStageKeypad/KeyStageKeypad.test.tsx
@@ -1,9 +1,10 @@
 import userEvent from "@testing-library/user-event";
 
-import renderWithTheme from "../../__tests__/__helpers__/renderWithTheme";
-import keyStageKeypad from "../../browser-lib/fixtures/keyStageKeypad";
-
 import KeyStageKeypad from "./KeyStageKeypad";
+
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+import keyStageKeypad from "@/browser-lib/fixtures/keyStageKeypad";
+
 
 const keyStages = ["1", "2", "3", "4"];
 const years = ["6", "7", "8", "9", "10", "11"];
