+++ src/components/SharedComponents/KeyStageKeypad/KeyStageKeypad.tsx
@@ -1,12 +1,11 @@
 import { FC } from "react";
 
-import { TeachersHomePageData } from "../../node-lib/curriculum-api";
-import OakLink from "../OakLink";
-import useAnalytics from "../../context/Analytics/useAnalytics";
-import type { KeyStageTitleValueType } from "../../browser-lib/avo/Avo";
-import useAnalyticsPageProps from "../../hooks/useAnalyticsPageProps";
-import useIsCurrent from "../MenuLinks/useIsCurrent";
-
+import { TeachersHomePageData } from "@/node-lib/curriculum-api";
+import OakLink from "@/components/OakLink";
+import useAnalytics from "@/context/Analytics/useAnalytics";
+import type { KeyStageTitleValueType } from "@/browser-lib/avo/Avo";
+import useAnalyticsPageProps from "@/hooks/useAnalyticsPageProps";
+import useIsCurrent from "@/components//MenuLinks/useIsCurrent";
 import BrushBorders from "@/components/SharedComponents/SpriteSheet/BrushSvgs/BrushBorders";
 import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import { Heading, UL, LI, P } from "@/components/SharedComponents/Typography";
