+++ src/pages/teachers/key-stages/[keyStageSlug]/subjects/index.tsx
@@ -16,7 +16,7 @@ import {
   KeyStageSubjectData,
 } from "@/node-lib/curriculum-api-2023/queries/subjectListing/subjectListing.schema";
 import getPageProps from "@/node-lib/getPageProps";
-import KeyStageKeypad from "@/components/KeyStageKeypad/KeyStageKeypad";
+import KeyStageKeypad from "@/components/SharedComponents/KeyStageKeypad";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import removeLegacySlugSuffix from "@/utils/slugModifiers/removeLegacySlugSuffix";
